* Bidi is supported.(FriBidi is required) * Arabic character shaping is supported. * ml_image module is cleaned up. * many minor bugs are fixed. * mlterm icons are added.(thanks to Kubota Tomohiro san.)

+++ src/ml_color_manager.h
@@ -27,7 +27,7 @@ typedef struct  ml_xcolor
 	XColor  xcolor ;
 #endif
 
-	int  is_loaded ;
+	int8_t  is_loaded ;
 	
 } ml_xcolor_t ;
